Introduction of an optional retry-mechanism for failed queries
Sometimes, execution of queries may fail for technical
reasons, often related to the stack getting overoaded. In these cases,
it might make sense to just wait a little and retry again.

The conditions for retry are narrowly defined, to avoid bluntly
retyring when exceptions are raised.
This definition migth evolve over time.

This is an optional feature, so the default behaviour is still the same.